requirements.txt
Flask
Flask_SocketIO
socketIO_client
flask-cors
squareup
bleak